+++ mongo_class.py
@@ -794,19 +794,11 @@ class MasterRep(Rep):
 
         super(MasterRep, self).__init__(name, user, passwd, host, port, auth,
                                         conf_file)
-        super(MasterRep, self).connect()
 
-        data = fetch_ismaster(self)
-
-        if data.get("ismaster"):
-            self.ismaster = data.get("ismaster")
-            self.issecondary = data.get("secondary")
-            self.repset = data.get("setName")
-            self.slaves = data.get("hosts", [])
-
-        else:
-            self.disconnect()
-            sys.exit("Error:  This is not a Master Replication server.")
+        self.ismaster = None
+        self.issecondary = None
+        self.repset = None
+        self.slaves = None
 
     def connect(self):